MM_30482 : Code clean up
