+++ types/database/database.d.ts
@@ -1,7 +1,11 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+import Model from '@nozbe/watermelondb/Model';
+
+// See LICENSE.txt for license information.
 import {Migration} from '@nozbe/watermelondb/Schema/migrations';
 import {AppSchema, Database} from '@nozbe/watermelondb';
+import {OperationType} from '../../app/database/admin/data_operator';
 import {DatabaseType} from '../../app/database/admin/database_manager';
 
 export type MigrationEvents = {
@@ -33,3 +37,19 @@ export type DefaultNewServer = {
 // A database connection is of type 'Database'; unless it fails to be initialize and in which case it becomes 'undefined'
 export type DBInstance = Database | undefined
 
+type RawApp = {
+    buildNumber: string,
+    createdAt: number,
+    id: string,
+    versionNumber: string,
+}
+
+type RecordValue = RawApp | any
+
+type DataFactory = {
+    db: Database,
+    generator? : (model: Model) => void,
+    optType?: OperationType,
+    tableName: string,
+    value: RecordValue,
+}
